spiral/requirements.txt
Django==1.5.1
MySQL-python==1.2.4
South==0.7.6
argparse==1.2.1
distribute==0.6.36
django-classy-tags==0.3.4.1
django-crispy-forms==1.2.5
mysql-connector-python==1.0.9
wsgiref==0.1.2
django-debug-toolbar==0.9.4
xlwt==0.7.5
